Converted terminal.py to cli.py and added improvements to select_directory()
